Factor out order validation tests
